+++ tools/psmisc/fuser/mount.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -e
+
+# fuser use stdin,stderr at the same time
+exec 2>&1
+
+fuser -mv /